adding purchases

+++ src/controllers/cart.controllers.js
@@ -1,9 +1,22 @@
 const catchError = require('../utils/catchError');
 const Cart = require('../models/Cart');
+const Product = require('../models/Product');
+const Category = require('../models/Category');
 
 const getAll = catchError(async(req, res) => {
+    const userId = req.user.id
     const results = await Cart.findAll({
-        where: { userId: 1}
+        where: { userId },
+        include:[
+            {
+                model: Product,
+                attributes: {exclude: [ "updatedAt", "createdAt", "id" ]},
+                include: {
+                    model: Category,
+                    attributes: ["name"]
+                }
+            }
+        ]
     });
     return res.json(results);
 });
@@ -20,16 +33,19 @@ const create = catchError(async(req, res) => {
 
 const remove = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.destroy({ where: {id} });
+    const userId = req.userId
+    const result = await Cart.destroy({ where: {id, userId} });
     if(!result) return res.sendStatus(404);
     return res.sendStatus(204);
 });
 
 const update = catchError(async(req, res) => {
+    const userlogged = req.user.id
     const { id } = req.params;
+    const { quantity } = req.body
     const result = await Cart.update(
-        req.body,
-        { where: {id}, returning: true }
+        { quantity },
+        { where: {id, userlogged}, returning: true }
     );
     if(result[0] === 0) return res.sendStatus(404);
     return res.json(result[1][0]);
